backend/requirements.txt
psutil>=5.9.0
websockets>=12.0
pystray>=0.19.5
Pillow>=10.0.0
wmi>=1.5.1
pywin32>=306
keyboard>=0.13.5
nvidia-ml-py>=12.0.0
